Renamed misago.permissions.edits to misago.permissions.postedits
